Refactor logging to work for VS Code and VS
This commit refactors a bunch of the logging that exists and tries to
fix the issue of NatvisDiagnostics not showing up in later debug
sessions.
