+++ src/DebugEngineHost.VSCode/HostLogger.cs
@@ -1,0 +1,46 @@
+﻿// // Copyright (c) Microsoft. All rights reserved.
+// // Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+
+namespace Microsoft.DebugEngineHost
+{
+    public static class HostLogger
+    {
+        private static HostLogChannel s_natvisLogChannel;
+        private static HostLogChannel s_engineLogChannel;
+
+        public static void EnableNatvisLogger(Action<string> callback)
+        {
+            if (s_natvisLogChannel == null)
+            {
+                // TODO: Support writing natvis logs to a file.
+                s_natvisLogChannel = new HostLogChannel(callback, null);
+            }
+        }
+
+        public static void EnableHostLogging(Action<string> callback, string logFile)
+        {
+            if (s_engineLogChannel == null)
+            {
+                s_engineLogChannel = new HostLogChannel(callback, logFile);
+            }
+        }
+
+        public static HostLogChannel GetEngineLogChannel()
+        {
+            return s_engineLogChannel;
+        }
+
+        public static HostLogChannel GetNatvisLogChannel()
+        {
+            return s_natvisLogChannel;
+        }
+
+        public static void Reset()
+        {
+            s_natvisLogChannel = null;
+            s_engineLogChannel = null;
+        }
+    }
+}
