+++ src/MIDebugEngine/Natvis.Impl/Natvis.cs
@@ -325,7 +325,7 @@ namespace Microsoft.MIDebugEngine.Natvis
                 XmlSerializer serializer = new XmlSerializer(typeof(AutoVisualizer));
                 if (!File.Exists(path))
                 {
-                    _process.Logger.GetNatvisChannelLog()?.WriteLine(ResourceStrings.FileNotFound, path);
+                    _process.Logger.NatvisLogger?.WriteLine(ResourceStrings.FileNotFound, path);
                     return false;
                 }
                 XmlReaderSettings settings = new XmlReaderSettings();
@@ -406,7 +406,7 @@ namespace Microsoft.MIDebugEngine.Natvis
             catch (Exception exception)
             {
                 // don't allow natvis failures to stop debugging
-                _process.Logger.GetNatvisChannelLog()?.WriteLine(ResourceStrings.ErrorReadingFile, exception.Message, path);
+                _process.Logger.NatvisLogger?.WriteLine(ResourceStrings.ErrorReadingFile, exception.Message, path);
                 return false;
             }
         }
@@ -451,7 +451,7 @@ namespace Microsoft.MIDebugEngine.Natvis
             {
                 // don't allow natvis to mess up debugging
                 // eat any exceptions and return the variable's value
-                _process.Logger.GetNatvisChannelLog()?.WriteLine("FormatDisplayString: " + e.Message);
+                _process.Logger.NatvisLogger?.WriteLine("FormatDisplayString: " + e.Message);
             }
             finally
             {
